Titlehandler for python.org `og:title`s at python.org are really broken (issue opened: python/pythondotorg#756), for example on [this page](https://www.python.org/dev/peps/pep-0448/): ```html <title>PEP 0448 -- Additional Unpacking Generalizations | Python.org</title> <meta property="og:title" content="Welcome to Python.org"> ```
So `og:title` isn't a suitable source for a title. Fixing it would require a lot of refactoring and since I don't have nearly any knowledge about Django templating I probably won't do it. Also few people probably are even interested in that amount of refactoring so I don't expect it being fixed soon.

In the meantime a titlehandler would be a suitable solution.